incorporated double tree, split up clustering some, and changed N_threads to num_threads
